ajustes varios para configuración y mejora en cobertura de codigo
