+++ Workers/AudioRemoval/tests/unit/adapters/message_handler_test.go
@@ -1,4 +1,4 @@
-package adapters
+package adapters_test
 
 import (
 	"encoding/json"
