+++ Workers/AudioRemoval/tests/unit/adapters/minio_test.go
@@ -1,4 +1,4 @@
-package adapters
+package adapters_test
 
 import (
 	"bytes"
