+++ Workers/AudioRemoval/tests/unit/adapters/rabbitmq_test.go
@@ -1,4 +1,4 @@
-package adapters
+package adapters_test
 
 import (
 	"testing"
