+++ Workers/AudioRemoval/tests/unit/adapters/storage_repository_test.go
