+++ Workers/AudioRemoval/tests/unit/domain/entities_test.go
@@ -1,4 +1,4 @@
-package domain
+package domain_test
 
 import (
 	"testing"
