+++ Workers/AudioRemoval/tests/unit/services/mp4_video_processing_test.go
@@ -1,4 +1,4 @@
-package services
+package services_test
 
 import (
 	"os"
