+++ nginx.conf
@@ -35,9 +35,9 @@ http {
     location / { proxy_pass http://api:8080/; }
   }
 
-  # MinIO Files → http://localhost:8081
+  # MinIO Files → http://localhost:8084
   server {
-    listen 8081;
+    listen 8084;
     client_max_body_size 500m;
     
     location /processed-videos/ {
